Run user run commands as an admin
Since user run can only be run by an admin, it can be useful to
enable admins to run commands with an admin privilege when
imperonsating users with user run.

For example, this enables admins to easily add networks to a user
past the max enable networks limit for that user.