Fix attempting to store last message ID with no backlog clients
Whenever we store a message in the message store; if this is
the first message we receive from the message target, we
store that message ID in our network delivered store. This is
used for later sending the backlog to clients not aware of
chathistory.

However, there was previously a bug where when that delievered
map was empty, i.e. when we had no clients not aware of chat
history; where we'd try on every message to store that ID,
but failed to do so (because StoreID was not called, because
ForEachClient never called its callback because there was no
client). We ended up querying the last message ID on every
inserted message, and not using it. Note that this is much
worse than when there was a client not aware of chat history:
we would only query and store that ID once, not on every
message.

With this bugfix, we skip querying the last message ID when
we know that will we not insert it. This saves a SELECT on
the Message + MessageTarget database on every message insert,
which could take up to a few seconds for large channels.