Avoid sending MARKREAD via Web Push for skipped previous notification
Commit d9aebc858f9f ("Wait 5 seconds for a MARKREAD before broadcasting
message Web push") has introduced a mechanism to skip Web Push
notifications for PRIVMSG if one of the clients replies with a
MARKREAD quickly enough. However we were still broadcasting the
MARKREAD message via Web Push in that case. Since there is no
notification to dismiss, skip it.

This requires bumping the timestamp only if we're actually going
to send a notification.

Ideally this last Web Push notification timestamp would be stored
in the DB instead of in-memory.