upstream: fix delay after connect commands
The rationale behind these is to add a delay until the NickServ
processes these commands which often authenticate the user to
successfully join channels that require the user to be authenticated
first. As such it must be ensured that the delay happens *before*
attempting to autojoin channels, which wasn't the case. Fix that.

Fixes: cef94ff662a6 ("upstream: sleep for a bit after sending connect commands")